L'utilisateur connecté via JAAS est reconnu comme stagiaire effectuant l'évaluation. Tout marche bien Next travail sur les relations evalution-stagiaire-session-formation

+++ src/fr/treeptik/dao/GenericDAO.java
@@ -36,7 +36,7 @@ public abstract class GenericDAO<T> {
 		return em.merge(entity);
 	}
 
-	public T findById(long entityID) {
+	public T findById(int entityID) {
 		return em.find(entityClass, entityID);
 	}
